Core: Merge Sean Cross's TtiCriteria patch. --- Notes, in no particular order:
I have added a property Criteria: TtiObject to TtiObjectList.  This is declared as TtiObject to get around circular reference issues.

I have also created a descendant TtiFilteredObjectList (in tiFilteredObjectList.pas) that just casts the Criteria back to TtiCriteria.

I have temporarily removed the Group by and Ordering functionality as it  isn't tested, and could break the existing implementation (Ordering shouldn't, but it isn't tested...). I will add this back in at a later stage.


I have added     procedure SelectRow(const ATableName : string; const AWhere : TtiQueryParams; const ACriteria: TPerCriteria); overload; virtual; abstract;

to TtiQuery.  This is not yet used in TtiQueryNonSql (and will give an error).  In TtiQuerySql, it is overridden to go the dirty work.  At the moment it uses Graeme's sql generation, but I want to change it to be parameterized.

This could be called in TVisDBAutoGenRead.OpenQuery to add Filtering to DbIndependant vistors in about 2 lines of code.  I will leave this as an exercise for the reader at the moment as I haven't time to write the unit tests.

I have added unit tests to Ado, BDE and IBX tests.  I have not run the IBX tests as I don't have interbase installed though.

I have a demo that I will upload to the binary groups shortly.

Todo in stage 2,3 etc:
Add filtering to DbIndependant visitors
Use paramaterization in sql generation
Add filtering to non sql automapping
Expand the range of datatypes supported, possible using variants. 



git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@655 68417047-fa15-0410-9c44-b939d5e0afe3