skips last character bugfix in stCharExistsL()
The pascal version of stCharExistsL (intiDataBuffer_Cli.pas) currently
always skips the last character of the string to search in. This also
means this function always returns False if the first parameter (the
string to search in) has a length of 1.

Thanks to "G.J. Doornink" <dkg@xs4all.nl> for the patch.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@1846 68417047-fa15-0410-9c44-b939d5e0afe3