LCL StringGrid Mediator column sizing fix
The old behaviour always resized the last column to fit in the
width of the grid (meaning no horizontal scrollbar ever). This
also ignored the developers preferred column width for the last
column.

The behaviour we actually wanted, was to only resize the last column
to fill the grid width, if the content of the grid was less than
the grid width. Last column must never be smaller than developers
preferred columnt width either. This is now so.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@1862 68417047-fa15-0410-9c44-b939d5e0afe3