LCL TListview mediator improvements.
It makes sure that the first listitem is selected in the listview as
the list is rebuilt or an item is added. If an item was selected
prior to rebuilding the list, it attempts to re-select it, falling
back to the first one.

If the list is well-constructed, it saves a click...
Thanks Michael van Canneyt for the patch.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@1984 68417047-fa15-0410-9c44-b939d5e0afe3