The code for TtiObjectVisitorSetObjectStateToDelete does seem to have worked.
Recursively setting the ObjectState using TtiObjectVisitorSetObjectStateToDelete visitor
did not seem to work at all. There was no unit tests for this either.

For now I commented the code, and simply called inherited, which seems to
fix this visitor. It works in my code too, but I still need to implement
a unit test for it to comfirm it 100%.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@1994 68417047-fa15-0410-9c44-b939d5e0afe3