new AParams parameter added to DatabaseExists(), CreateDatabase() and DropDatabase()
Thanks to Brave Cobra <bravecobra@gmail.com> for the contribution.

Added a patch for MSSQL. We need to a connect to a SQL Server
database in mixed mode (as a DB user, not a Windows user which is the
only way at the moment)

I (Graeme) modified his patch and changed AParams to a constant parameter -
as is done for Connect(). I also implemented the change for all the
remaining persistence layers.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2031 68417047-fa15-0410-9c44-b939d5e0afe3