Not all supported persistence layers support opening non-returning SQL statements
In some persistence layers we need to call ExecSQL for SQL statements
than don't return a resultset. eg: Update, Insert or Delete etc.

I'm not sure what is the best way to do this. For now I am simply
checking for the text "select" - this could probably be improved
on.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2116 68417047-fa15-0410-9c44-b939d5e0afe3