MGM for LCL: StingGrid header alignment change.
The text alignment of the StringGrid header is now default to centred.
This is the same as what fpGUI does, and how spreadsheet applications
work too.

This is independent of the alignment of the data of course. Data alignment
is configurable by the application developer using the field info text.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2151 68417047-fa15-0410-9c44-b939d5e0afe3