TtiObject.Equals now has a new parameter ADeepCheck
Originally the .Equals() call did a deep comparison checking every property when it does
a comparison. This is an CPU expensive task.

The default behaviour of .Equals() have now changed to do a "shallow" comparison, by
only looking at the OID values of the to objects. This obviously means it is a much quicker
comparison, but not as thorough. By tiOPF's definition, the OID is a unique
identifier of an object. So if two objects have the same OID values, for all intense purposes,
they are equal. So this behaviour should still work with all existing code.

Now if you still want the "deep compasion", then simply pass the value True as the second
parameter to .Equals(), and you'll get the same behaviour as before.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2199 68417047-fa15-0410-9c44-b939d5e0afe3