tests: new ZEOS+Firebird testing unit.
The tests indicate that the majority of Zeos+Firebird persistence layer is working, but there
are still quite a few tests that fail.

NOTE:
Rememeber to add
    Params=protocol=firebird-xxx
to your unit test ini configuration file, so the Zeos+Firebird persistence layer is correctly
loaded. If you don't, you will get the 'protocol is not specified' error.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2216 68417047-fa15-0410-9c44-b939d5e0afe3