Revert tiPad0 change from r2155.
I had to dig around to make sure.
Looking at the expected value in the unit tests, and the expected
behaviour of tiPad0 elsewhere in the tiOPF framework, tiPad0 should
behave slightly different to tiPadL. So internally calling tiPadL was wrong.

tiPadL: Expects to work with strings. Padding is added to the left, and
characters are removed from the left.

tiPad0: Expects to work with numbers in string format. Padding is added
to the left, and characters are removed from the _right_.

As I said, the unit tests for tiPad0 proves this, and so does the usage
of tiPad0 in the TtiFieldCurrency.GetAsString.

It's not 100% clear-cut what the intent of tiPad0 was in cases where the AValue is longer
than the ALen value. But for now we go with how the unit tests was written and what the
tests expected as results.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF3/Trunk@2243 68417047-fa15-0410-9c44-b939d5e0afe3