Auto-gen delete class is extended to support custom primary key fields.
Currently, on Delete visitor, pk fieldname is hardcoded to OID. That is a problem while
trying to work with legacy databases. This patch simply adds a property 'OIDFieldName'
on TVisDBAutoGenDelete where developer can specify fieldname of pk to be removed.
To ensure backwars compatibility, OIDFiledName has a default value of 'OID'.

             from Raul Ferriz <raul.ferriz@gmail.com>

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF3/Trunk@2254 68417047-fa15-0410-9c44-b939d5e0afe3