Fixes a NotifyObservers bug introduced in r1376.
Even when we add, remove, extract etc objects from a objectlist, the
Subject in the NotifyObservers call should always be the objectlist,
not the item being added, removed etc. This is as per the Observer
documentation in the GoF book.

I added some new unit tests to confirm this change.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2303 68417047-fa15-0410-9c44-b939d5e0afe3