Cleaned up Semaphore related code in tiPool.
The semaphore code had lots of IFDEF's. What made matters worse, is that
each OS seems to have a different way of doing this. Not even Linux
and FreeBSD was the same. Also the FPC+FreeBSD+Semaphore support in FPC
is broken.

So I implemented a basic cross-platform and cross-compiler semaphore
class. So far this works just fine, and all tiPool unit tests pass on
Windows, Linux and FreeBSD.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2309 68417047-fa15-0410-9c44-b939d5e0afe3