Observer method signature change to support Data object
The list mediator classes expected the affected Data object to be passed
in via the Update() method. But since the Observer design pattern was
fixed in tiObject, that was not the case for the list mediators any more.

So we extended the Observer implementation by including the affected
data object as a new parameter. Now the list mediators and any other
custome mediators have easy access to the data object that was added or
deleted from a list.

With this commit, I have tested the LCL and fpGUI mediators. Next up, I'll
update the VCL mediators too.

git-svn-id: https://tiopf.svn.sourceforge.net/svnroot/tiopf/tiOPF2/Trunk@2321 68417047-fa15-0410-9c44-b939d5e0afe3