Adds new ObjectState posLoading
The idea is to have StartLoading, StopLoading and IsLoading methods in
your business classes. So when reading in a object from the DB, the
setter methods of properties can check the IsLoading values and determain
what it must do. eg: while loading on object you might not whant to
notify observers, or check permission to set a property etc.

The idea was suggestion by Michael van Canneyt

Merged from tiOPF2 commit 5ab48a8