Updated usage requirements for OIDInteger and OIDInt64
I also mentioned the initialisation of the seed value must not be zero,
because that is a special case to allow code like "if OID.IsNull then..."

Some years ago, any seed value could be used, but about 3-4 years ago,
IsNull() support was implemented, which changed the rules slightly. Unit
Tests must still be implemented to reflect this behaviour.