Removed tiDataset from the tiOPFLCL.lpk package.
Something in Lazarus changed and now when FPC compiles the package we get
units in the Core package recompiled for the tiOPFLCL package. The latter
uses a separate unit output directory. In turn FPC detects a difference or
conflicting *.ppu files and then fails to compile the package.

Removing the Options directory for the tiOPFLCL search path and removing
the tiDataset unit resolves the problem.

NOTE: This is only a Lazarus issue.