Moved Demo 21's 'displayhelpers.pas' unit to Core/tiDisplayHelpers.pas
There are multiple copies of this unit in the Demos directory. I also have
my own copies in various personal projects. It makes sense to have this
as a shared unit.

See the unit header comment for details of what it does and what it can
be used for. In summary, if you use mediators the display helpers can do
GUI formatting of properties, so the BOM doesn't have too.