fpgui: listbox list mediator needs a custom CreateSubMediator implementation.
We got a Index out of Bounds error when BOM objects got marked as deleted.
This was because we used the index value of all available objects, and
mapped that directly to the GUI component. The index should rather be
only of the objects we actually are going to display.

Demo 20 explains this best.